feat: bbs 리팩토링, cors 설정

+++ src/main/java/com/toyproject/jobadream/bbs/dto/CommentDto.java
@@ -9,5 +9,5 @@ public class CommentDto {
 	private Integer id;
 	private String comment;
 	private LocalDateTime commentDate;
-	private Long postId;
+	private Integer postId;
 }
